move forward_msg_cache, forward_msg_queue
