+++ lib/tests/streamlit/runtime/forward_msg_cache_test.py
@@ -19,9 +19,9 @@ import unittest
 
 from streamlit import config, RootContainer
 from streamlit.runtime import app_session
-from streamlit.forward_msg_cache import ForwardMsgCache
-from streamlit.forward_msg_cache import create_reference_msg
-from streamlit.forward_msg_cache import populate_hash_if_needed
+from streamlit.runtime.forward_msg_cache import ForwardMsgCache
+from streamlit.runtime.forward_msg_cache import create_reference_msg
+from streamlit.runtime.forward_msg_cache import populate_hash_if_needed
 from streamlit.elements import legacy_data_frame as data_frame
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.stats import CacheStat
