+++ lib/tests/streamlit/runtime/forward_msg_queue_test.py
@@ -22,7 +22,7 @@ from parameterized import parameterized
 
 from streamlit import RootContainer
 from streamlit.cursor import make_delta_path
-from streamlit.forward_msg_queue import ForwardMsgQueue
+from streamlit.runtime.forward_msg_queue import ForwardMsgQueue
 from streamlit.elements import legacy_data_frame
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
