+++ lib/tests/streamlit/scriptrunner/script_runner_test.py
@@ -37,7 +37,7 @@ from streamlit.scriptrunner.script_requests import (
     ScriptRequests,
 )
 from streamlit.session_data import SessionData
-from streamlit.forward_msg_queue import ForwardMsgQueue
+from streamlit.runtime.forward_msg_queue import ForwardMsgQueue
 from streamlit.scriptrunner import (
     ScriptRunner,
     ScriptRunnerEvent,
@@ -46,7 +46,7 @@ from streamlit.scriptrunner import (
     StopException,
 )
 from streamlit import source_util
-from streamlit.state.session_state import SessionState, WidgetMetadata
+from streamlit.state.session_state import SessionState
 from streamlit.uploaded_file_manager import UploadedFileManager
 from tests import testutil
