+++ lib/tests/testutil.py
@@ -21,7 +21,7 @@ from unittest.mock import patch
 
 from streamlit import config
 from streamlit.runtime.app_session import AppSession
-from streamlit.forward_msg_queue import ForwardMsgQueue
+from streamlit.runtime.forward_msg_queue import ForwardMsgQueue
 from streamlit.proto.Delta_pb2 import Delta
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.scriptrunner import (
